requirements.txt
fastapi
uvicorn
gunicorn
google-api-python-client
google-auth-oauthlib
python-dotenv
sentence-transformers
pandas
python-docx
torch
faiss-cpu
pymupdf
google-cloud-storage